+++ web/search/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+# Application Configuration
+
+
+class SearchConfig(AppConfig):
+    name = 'search'
