+++ web/search/views.py
@@ -1,0 +1,48 @@
+from django.contrib.postgres.search import SearchVector
+from django.contrib.postgres.search import SearchQuery
+from django.contrib.postgres.search import SearchRank
+
+from django.views.generic.list import ListView
+from django.core.paginator import Paginator
+from django.conf import settings
+
+from .forms import SearchForm
+
+from talks.models import Talk
+
+
+class SearchTalksView(ListView):
+    model = Talk
+    template_name = 'search.html'
+    paginate_by = settings.PAGE_SIZE
+
+    def _search_talks(self, q):
+        vector = SearchVector('title', weight='A') + SearchVector('description', weight='B')
+        query = SearchQuery(q)
+        rank = SearchRank(vector, query)
+        search_results = Talk.published_objects.annotate(rank=rank)
+        # Filter by minimum rank
+        search_results = search_results.filter(rank__gte=0.1).distinct()
+        # Sort by rank (descendant)
+        search_results = search_results.order_by('-rank', '-created')
+        return search_results
+
+    def get_context_data(self, **kwargs):
+        context = super(SearchTalksView, self).get_context_data(**kwargs)
+
+        search_form = SearchForm(self.request.GET)
+        context['search_form'] = search_form
+
+        if search_form.is_valid():
+            q = search_form.cleaned_data['q']
+            context['search_query'] = q
+
+            search_results = self._search_talks(q)
+            paginator = Paginator(search_results, self.paginate_by)
+
+            page = 1
+            if "page" in self.request.GET:
+                page = self.request.GET["page"]
+            context['object_list'] = paginator.get_page(page)
+
+        return context
