+++ web/settings/base.py
@@ -50,6 +50,7 @@ INSTALLED_APPS = [
     'playlists',
     'channels',
     'talks',
+    'search',
 ]
 
 MIDDLEWARE = [
